src/main/java/project/altynbaev/dictionary/country/controller/CountryController.java
package project.altynbaev.dictionary.country.controller;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import project.altynbaev.dictionary.country.service.CountryService;
import project.altynbaev.dictionary.country.view.CountryView;

import java.util.List;

@Api(value = "CountryController")
@RestController
@RequestMapping("/api/countries")
public class CountryController {
    private final CountryService countryService;

    @Autowired
    public CountryController(CountryService countryService) {
        this.countryService = countryService;
    }

    @ApiOperation(value = "Получить список стран из справочника", httpMethod = "GET")
    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "Success", response = CountryView.class)
    })
    @GetMapping()
    public List<CountryView> getList() {
        return countryService.getList();
    }
}
